adding trace info as common attributes, refactoring options variable and startNewSession
