+++ packages/crashlytics/src/api.ts
@@ -20,7 +20,6 @@ import { CRASHLYTICS_TYPE } from './constants';
 import { Crashlytics, CrashlyticsOptions } from './public-types';
 import { Provider } from '@firebase/component';
 import { AnyValueMap, SeverityNumber } from '@opentelemetry/api-logs';
-import { trace } from '@opentelemetry/api';
 import { CrashlyticsService } from './service';
 import { flush, setCommonLogAttributes } from './helpers';
 import { CrashlyticsInternal } from './types';
@@ -104,18 +103,6 @@ export function recordError(
   if (frameworkAttributesProvider) {
     const frameworkAttributes = frameworkAttributesProvider();
     Object.assign(customAttributes, frameworkAttributes);
-  }
-
-  // Add trace metadata
-  const activeSpanContext = trace.getActiveSpan()?.spanContext();
-  if (crashlytics.app.options.projectId && activeSpanContext?.traceId) {
-    customAttributes[
-      'logging.googleapis.com/trace'
-    ] = `projects/${crashlytics.app.options.projectId}/traces/${activeSpanContext.traceId}`;
-    if (activeSpanContext?.spanId) {
-      customAttributes['logging.googleapis.com/spanId'] =
-        activeSpanContext.spanId;
-    }
   }
 
   setCommonLogAttributes(crashlytics, customAttributes);
